requirements.txt
cffi==1.16.0
colorama==0.4.6
cryptography==42.0.8
iniconfig==2.0.0
packaging==24.1
pillow==10.3.0
pluggy==1.5.0
pycparser==2.22
pytest==8.2.2
